initialization

+++ src/zendron/__main__.py
@@ -1,29 +1,37 @@
+import argparse
 import logging
+import os
+import os.path as osp
 import subprocess
 
-from zendron import load
+import hydra
+from omegaconf import DictConfig
+
+from zendron import init, load
 
 log = logging.getLogger(__name__)
 
 
-def main():
-    log.info("Sync Starting")
-    load.main()
-    subprocess.run("dendron importPod --podId dendron.markdown --wsRoot .", shell=True)
-    subprocess.run("rm -r zotero_pod", shell=True)
-    log.info("Sync Complete")
+@hydra.main(
+    version_base=None, config_path=osp.join(os.getcwd(), "conf"), config_name="config"
+)
+def main(cfg: DictConfig) -> None:
+    init.main()
+    if cfg.remove:
+        from zendron import remove
+
+        log.info("Remove Starting")
+
+        remove.main(cfg)
+    else:
+        log.info("Sync Starting")
+        load.main(cfg)
+        subprocess.run(
+            "dendron importPod --podId dendron.markdown --wsRoot .", shell=True
+        )
+        subprocess.run("rm -r zotero_pod", shell=True)
+        log.info("Sync Complete")
 
 
 if __name__ == "__main__":
     main()
-
-
-import shutil
-
-directory_to_remove = "/path/to/directory"
-
-try:
-    shutil.rmtree(directory_to_remove)
-    print(f"Directory '{directory_to_remove}' and its contents have been removed.")
-except OSError as e:
-    print(f"Error: {e.filename} - {e.strerror}.")
